fix test for valid moves
fixed the valid moves to so that it test the valid moves from every square with the pieces in their initial locations. I wanted to add the tests to the game logic test class but they failed while in there so I kept them as their own class. In the future they should probably be moved to be with the other game logic tests.