Update getValidMoves to follow copenhagen rules
I fixed the getValidMoves method to allow all pieces to pass over the center square but only the king to land on it. Previously, the king could pass over the center square but other pieces could not. The king is still the only piece that can land on the center square.