Fix king capture around throne
This fixes the rules by which the king is captured. Before the king
could be captured by being surrounded on an edge but after re-reading
the rules I realized that this does not capture the king. I changed it
so the king is not captured when surrounded on an edge but the king IS
captured when it is adjacent to the throne and surrounded