save prompt only appears now if not game not saved
Changed code so now the save prompt that appears when exiting a game
only appears when the game has not already been save or if no moves have
been made. I also switched the concede button to just go to final menu
since before it asks to save game when not necessary and closed without
letting the user play again.