Correctly Display the Number of Pieces Left
This now display the number of pieces left for both sides in all cases.
However, there is a bug not related to this that I noticed.

DEFECT: loading a saved game and then starting up a new game does not
create new timers and the game will end prematurely. What is displayed
and the timer actually is does not match up.