Fix load clock bug
This fixes a problem where the clocks would not be reset properly when
a game is loaded from file. It would display the correct time but that
was not the same time that was being tracked. So a player would be
declared winner in the newly loaded game when the timer from the
previous game expired