add draw functionality
This adds the ability for the game to end in a draw. This will happen
if no pieces are captured after 50 moves.
I also changed the final menu a bit, so now it displays the winning
side’s piece if there is a winner or explains why there was a draw.