Correct turn counting
Fixed the way it counts turns. Originally it counted each players turn
as 1. It now, correctly, increments the turn count only after both
players have played. The total number of moves is still saved in the
save file and used to keep track of the turn count. But this total
number of moves is never displayed to the user.